Add docs
and reorganize existing stub pages
