--- Letterbook.Docs/Pages/Architecture/Index.cshtml
@@ -1,23 +1,0 @@
-@page "/architecture/"
-@model Letterbook.Docs.Pages.Architecture.Index
-
-@{
-ViewData["Title"] = "Letterbook Architecture";
-ViewData["Brand"] = "Architecture";
-}
-
-@using Letterbook.Docs.Markdown
-@implements IRenderStatic<Index>
-
-@functions {
-    public List<Index> GetStaticProps(RenderContext ctx)
-    {
-        var docs = ctx.Resolve<LoaderFactory>();
-        return [new Index(docs)];
-    }
-}
-
-@foreach (var doc in Model.Docs.GetAll<MarkdownAdr>())
-{
-    
-}
