--- Letterbook.Docs/Pages/Architecture/Page.cshtml
@@ -1,19 +1,0 @@
-@page
-@model Letterbook.Docs.Pages.Architecture.Page
-
-@{
-    Layout = null;
-}
-
-<!DOCTYPE html>
-
-<html>
-<head>
-    <title></title>
-</head>
-<body>
-<div>
-    
-</div>
-</body>
-</html>
