--- Letterbook.Docs/Pages/Architecture/Page.cshtml.cs
@@ -1,11 +1,0 @@
-using Microsoft.AspNetCore.Mvc.RazorPages;
-
-namespace Letterbook.Docs.Pages.Architecture;
-
-public class Page : PageModel
-{
-	public void OnGet()
-	{
-		
-	}
-}
