+++ Letterbook.Docs/Pages/Blog/Index.cshtml
@@ -3,8 +3,7 @@
 @model Letterbook.Docs.Pages.Blog.Index
 
 @{
-    ViewData["Title"] = "Letterbook Blog";
-    ViewData["Brand"] = "Blog";
+    ViewData["Title"] = "Blog";
 }
 
 @implements IRenderStatic<Index>
