+++ Letterbook.Docs/Pages/Docs/Index.cshtml
@@ -1,25 +1,22 @@
 @page "/docs/"
 @model Letterbook.Docs.Pages.Docs.Index
+
 @{
-    ViewData["Title"] = "Docs";
+ViewData["Title"] = "Architecture";
 }
 
 @using Letterbook.Docs.Markdown
 @implements IRenderStatic<Index>
 
 @functions {
-
     public List<Index> GetStaticProps(RenderContext ctx)
     {
-        var categories = ctx.Resolve<Markdown.LoaderFactory>();
-        return [new Index(categories)];
+        var docs = ctx.Resolve<LoaderFactory>();
+        return [new Index(docs)];
     }
-
 }
 
-<ul>
-@foreach (var p in Model.Categories.GetAll<MarkdownCategory>())
+@foreach (var doc in Model.Docs.GetAll<MarkdownAdr>())
 {
-    <li><a href=@($"docs/{p.Category ?? "page"}/{p.Slug}")>@p.Slug</a></li>
+    
 }
-</ul>
