+++ Letterbook.Docs/Pages/Docs/Page.cshtml
@@ -1,37 +1,19 @@
-@page "/docs/{Category}/{Slug}"
-@model Letterbook.Docs.Pages.Docs.Page
-
-@using Letterbook.Docs.Markdown
-@implements IRenderStatic<Page>
-@functions
-{
-    public List<Page> GetStaticProps(RenderContext ctx)
-    {
-        var docs = ctx.Resolve<Markdown.LoaderFactory>();
-        return docs.LoadFrom<Markdown.LoadCategories, MarkdownCategory>("_pages").GetAll<MarkdownCategory>()
-            .Where(each => each.Category != null && each.Slug != "index")
-            .OrderBy(each => each.Category ?? each.Slug)
-            .ThenBy(each => each.Order)
-            .ThenBy(each => each.Slug)
-            .Select(each => new Page(docs)
-            {
-                Category = each.Category?.ToLowerInvariant() ?? "page",
-                Slug = each.Slug
-            }).ToList();
-    }
-
-}
+@page
+@model Page
 
 @{
-    if (Model.Source == null)
-    {
-        Response.Redirect("/404");
-        return;
-    }
-
-
-    Response.StatusCode = 200;
-    ViewData["Title"] = Model.Source.Title;
+    Layout = null;
 }
 
-@Model.Html
+<!DOCTYPE html>
+
+<html>
+<head>
+    <title></title>
+</head>
+<body>
+<div>
+    
+</div>
+</body>
+</html>
