+++ Letterbook.Docs/Pages/Guides/Index.cshtml
@@ -1,0 +1,25 @@
+@page "/guides/"
+@model Index
+@{
+    ViewData["Title"] = "Guides";
+}
+
+@using Letterbook.Docs.Markdown
+@implements IRenderStatic<Index>
+
+@functions {
+
+    public List<Index> GetStaticProps(RenderContext ctx)
+    {
+        var categories = ctx.Resolve<Markdown.LoaderFactory>();
+        return [new Index(categories)];
+    }
+
+}
+
+<ul>
+@foreach (var p in Model.Categories.GetAll<MarkdownCategory>())
+{
+    <li><a href=@($"docs/{p.Category ?? "page"}/{p.Slug}")>@p.Slug</a></li>
+}
+</ul>
