+++ Letterbook.Docs/Pages/Guides/Index.cshtml.cs
@@ -1,0 +1,10 @@
+using Letterbook.Docs.Markdown;
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Mvc.RazorPages;
+
+namespace Letterbook.Docs.Pages.Guides;
+
+public class Index([FromServices] LoaderFactory categories) : PageModel
+{
+	public LoadCategories Categories { get; set; } = categories.LoadFrom<LoadCategories, MarkdownCategory>("_pages");
+}
