+++ Letterbook.Docs/Pages/Shared/Nav.cshtml
@@ -13,13 +13,14 @@
     {
         ("/", "Home"),
         ("/blog", "Blog"),
+        ("/docs", "Documentation"),
     };
 }
 
 <nav class="top-nav">
     <ul class="align-bottom">
         <li><a href="https://letterbook.com"><img href="/img/letterbook.png" alt="Letterbook logo" src="/img/letterbook.svg" class="logo"/></a></li>
-        <li><h1 class="letterbook-brand">@ViewData["Title"]</h1></li>
+        <li><h1 class="letterbook-brand">@ViewData["Brand"]</h1></li>
     </ul>
     <ul class="align-bottom">
         <li>
