+++ Letterbook.Docs/Pages/Shared/_Layout.cshtml
@@ -1,4 +1,9 @@
 ﻿@inherits Microsoft.AspNetCore.Mvc.Razor.RazorPage<TModel>
+
+@{
+ViewData["Brand"] = "Letterbook Project";
+}
+
 <!doctype html>
 <html lang="en">
 <head>
